requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
asn1crypto==0.24.0        # via cryptography
certifi==2018.8.24        # via requests
cffi==1.11.5              # via cryptography
chardet==3.0.4            # via requests
click==7.0                # via pip-tools, rq
croniter==0.3.25          # via rq-scheduler
cryptography==2.4.2       # via -r requirements.in
django-cors-headers==3.2.0  # via -r requirements.in
django-extensions==2.1.3  # via -r requirements.in
django-rq==1.2.0          # via -r requirements.in
django==2.2.9             # via -r requirements.in, django-cors-headers, django-rq, djangorestframework
djangorestframework-jwt==1.11.0  # via -r requirements.in
djangorestframework==3.11.0  # via -r requirements.in
flake8-commas==2.0.0      # via -r requirements.in
flake8==3.5.0             # via -r requirements.in, flake8-commas
gunicorn==19.9.0          # via -r requirements.in
idna==2.7                 # via cryptography, requests
isort==4.3.4              # via -r requirements.in
mccabe==0.6.1             # via flake8
pip-tools==5.3.1          # via -r requirements.in
psycopg2==2.8.3           # via -r requirements.in
pycodestyle==2.3.1        # via flake8
pycparser==2.19           # via cffi
pyflakes==1.6.0           # via flake8
pyjwt==1.7.1              # via -r requirements.in, djangorestframework-jwt
python-dateutil==2.7.3    # via croniter
python-dotenv==0.10.1     # via -r requirements.in
pytz==2018.5              # via django
redis==2.10.6             # via rq
requests-mock==1.5.2      # via -r requirements.in
requests==2.19.1          # via -r requirements.in, requests-mock
rq-scheduler==0.8.3       # via -r requirements.in
rq==0.12.0                # via django-rq, rq-scheduler
six==1.11.0               # via cryptography, django-extensions, pip-tools, python-dateutil, requests-mock
sqlparse==0.3.0           # via django
urllib3==1.23             # via requests
werkzeug==0.14.1          # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
